+++ worker/gitops_test.go
@@ -539,36 +539,3 @@ func Test_cleanupTrigger(t *testing.T) {
 	})
 	assert.False(t, triggered, "Should not trigger on missing app")
 }
-
-func Test_cueEnvironmentsToManifests(t *testing.T) {
-
-	const cueTemplate = `
-import "text/template"
-
-_instances: [
-	"first",
-	"second",
-]
-
-configs: [ for instance in _instances {
-	app:       template.Execute("myapp-{{ . }}", instance)
-	env:       "production"
-	namespace: "production"
-	chart: {
-	repository: "https://chart.onechart.dev"
-	name:       "cron-job"
-	version:    0.32
-	}
-	values: {
-	image: {
-		repository: "<account>.dkr.ecr.eu-west-1.amazonaws.com/myapp"
-		tag:        "1.1.1"
-	}
-	}
-}]
-`
-
-	manifests, err := cueEnvironmentsToManifests([]string{cueTemplate})
-	assert.Nil(t, err)
-	assert.Equal(t, 2, len(manifests))
-}
